% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SieveSGDFunction.R
\name{clean_up_result}
\alias{clean_up_result}
\title{Clean up the fitted model}
\usage{
clean_up_result(sieve.model)
}
\arguments{
\item{sieve.model}{a sieve sgd model.}
}
\value{
a processed sieve.model, adding function names and extract the best model
}
\description{
Clean up the fitted model
}
