% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clrCounts1.R
\docType{data}
\name{clrCounts1}
\alias{clrCounts1}
\title{Simulated CLR-transformed count table}
\format{
A data frame containing CLR-transformed counts, with 200 genes and 500 samples.
}
\usage{
clrCounts1
}
\description{
A simulated CLR-transformed count table with genes on the rows and samples on the columns for one condition.
}
\keyword{datasets}
