% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clrCounts2.R
\docType{data}
\name{clrCounts2}
\alias{clrCounts2}
\title{Simulated DV genes CLR-transformed count table}
\format{
A data frame containing CLR-transformed counts, with 500 genes and 200 samples per group.
}
\usage{
clrCounts2
}
\description{
A simulated CLR-transformed count table with genes on the rows and samples on the columns for two groups. Ten percent of the genes are DV genes (gene1 to gene50).
}
\keyword{datasets}
