% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clrCounts3.R
\docType{data}
\name{clrCounts3}
\alias{clrCounts3}
\title{Simulated DE genes CLR-transformed count table}
\format{
A data frame containing CLR-transformed counts, with 500 genes and 200 samples per group.
}
\usage{
clrCounts3
}
\description{
A simulated CLR-transformed count table with genes on the rows and samples on the columns for two groups. Ten percent of the genes are DE genes (gene1 to gene50).
}
\keyword{datasets}
