% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clrDV.R
\name{clrDV}
\alias{clrDV}
\title{Differential variability (DV) test of CLR-transformed RNA-Seq data}
\usage{
clrDV(data = NULL, group = NULL)
}
\arguments{
\item{data}{A CLR-transformed count matrix with genes/transcripts on the rows and
samples on the columns.}

\item{group}{A vector specifying the group labels of the data.}
}
\value{
An object of `\code{clrDV}' class that contains the results of the DV test and
associated information:
\item{DV}{The difference of standard deviation (\code{sigma}) between
group 2 and group 1 (\code{sigma2} \code{-} \code{sigma1}).}
\item{se}{The standard error of \code{DV}.}
\item{z}{The observed Wald statistic.}
\item{pval}{The unadjusted p-value of Wald test.}
\item{adj_pval}{The p-value of the Wald test adjusted using the Benjamini-Yekutieli procedure.}
\item{sigma1}{The maximum likelihood estimate of the standard deviation parameter for group 1.}
\item{se.sigma1}{The standard error of the maximum likelihood estimate of \code{sigma1}.}
\item{z.sigma1}{The Wald statistic for \code{sigma1}.}
\item{p.sigma1}{The p-value of the Wald test for \code{sigma1}.}
\item{sigma2}{The maximum likelihood estimate of the standard deviation parameter for group 2.}
\item{se.sigma2}{The standard error of the maximum likelihood estimate of \code{gamma2}.}
\item{z.sigma2}{The Wald statistic for \code{sigma2}.}
\item{p.sigma2}{The p-value of the Wald test for \code{sigma2}.}
}
\description{
Model CLR-transformed RNA-Seq data using the skew-normal distribution,
and then conduct a statistical test for finding genes/transcripts with
differential variability using the Wald test.
}
\examples{
   library(SIEVE)
   data(clrCounts2) # first 50 genes (gene1 to gene50) are DV genes
   groups <- c(rep(0, 200), rep(1, 200))
   clrDV_test <- clrDV(clrCounts2, group = groups)
   sum(is.na(clrDV_test))  # check NA values
   head(clrDV_test, 5)  # adj_pval < 0.05, DV genes
   tail(clrDV_test, 5)  # adj_pval > 0.05, non-DV genes


}
