% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clrSIEVE.R
\name{clrSIEVE}
\alias{clrSIEVE}
\title{Sieve Differential Expression, Variability and Skewness Genes
for RNA-Seq Data Based on Skew-normal Distribution}
\usage{
clrSIEVE(
  clrSeq_result = NULL,
  alpha_level = 0.05,
  order_DE = FALSE,
  order_LFC = FALSE,
  order_DS = FALSE,
  order_sieve = FALSE
)
}
\arguments{
\item{clrSeq_result}{The result of clrSeq() function.}

\item{alpha_level}{The adjusted p-value significance cutoff used for flagging
the differential expression, variability and skewness genes.}

\item{order_DE}{Logical string, "FALSE" for no ordering, "TRUE" for ordering by the value of \code{DE}.}

\item{order_LFC}{Logical string, "FALSE" for no ordering, "TRUE" for ordering by the value of \code{LFC}.}

\item{order_DS}{Logical string, "FALSE" for no ordering, "TRUE" for ordering by the value of \code{DS}.}

\item{order_sieve}{Character/logical string specifying the order method. Possibilities are "DE"
the value of \code{DE}, "LFC" for log2-fold-change of variability,
"DV" for the value of \code{DV} or "FALSE" for no ordering.}
}
\value{
Lists of \code{clrDE_test}, \code{clrDV_test}, \code{clrDS_test} and
\code{clrSIEVE_tests} classes that contain the results of the DE, DV,
DS tests and one-stop those three tests, and associated information:
\item{clrDE_test}{A data.frame contating the results of differential expression test.}
\item{clrDV_test}{A data.frame contating the results of differential variability test.}
\item{clrDS_test}{A data.frame contating the results of differential skewness test.}
\item{clrSIEVE_tests}{A data.frame contating the results of differential expression, variability and skewness tests.}
\item{DE}{The difference of mean (\code{mu}) between group 2 and group 1 (\code{mu2} \code{-} \code{mu1}).}
\item{se_DE}{The standard error of \code{DE}.}
\item{z_DE}{The observed Wald statistic of \code{DE}.}
\item{pval_DE}{The unadjusted p-value of Wald test of \code{DE}.}
\item{adj_pval_DE}{The p-value of the Wald test of  adjusted using the Benjamini-Yekutieli procedure.}
\item{mu1}{The maximum likelihood estimate of the mean parameter for group 1.}
\item{mu2}{The maximum likelihood estimate of the mean parameter for group 2.}
\item{de_indicator}{1: DE gene, 0: non-DE gene.}
\item{SD_ratio}{The ratio of standard deviation (\code{sigma}) between group 2 and group 1 (\code{sigma2}\code{/}\code{sigma1}.). }
\item{LFC}{\code{log2} fold change: \code{log2}(\code{sigma2}\code{/}\code{sigma1}).}
\item{DV}{The difference of standard deviation (\code{sigma}) between group 2 and group 1 (\code{sigma2} \code{-} \code{sigma1}.).}
\item{se_DV}{The standard error of \code{DV}.}
\item{z_DV}{The observed Wald statistic of \code{DV}.}
\item{pval_DV}{The unadjusted p-value of Wald test of \code{DV}.}
\item{adj_pval_DV}{he p-value of the Wald test of \code{DV} adjusted using the Benjamini-Yekutieli procedure.}
\item{sigma1}{The maximum likelihood estimate of the standard deviation parameter for group 1.}
\item{sigma2}{The maximum likelihood estimate of the standard deviation parameter for group 2.}
\item{dv_indicator}{1: DV gene, 0: non-DV gene.}
\item{DS}{The difference of skewness parameter (\code{gamma}) between group 2 and group 1 (\code{gamma2} \code{-} \code{gamma1}.).}
\item{se_DS}{The standard error of \code{DS}.}
\item{z_DS}{The observed Wald statistic of \code{DS}.}
\item{pval_DS}{The unadjusted p-value of Wald test of \code{DS}.}
\item{adj_pval_DS}{The p-value of the Wald test of \code{DS} adjusted using the Benjamini-Yekutieli procedure.}
\item{gamma1}{The maxiimum likelihood estimate of the skewness parameter for group 1.}
\item{gamma2}{The maxiimum likelihood estimate of the skewness parameter for group 2.}
\item{ds_indicator}{1: DS gene, 0: non-DS gene.}
}
\description{
Model CLR-transformed RNA-Seq data using the skew-normal
distribution, and then conduct statistical tests for finding
genes/transcripts with differential expression, variability
and skewness using the Wald test.
}
\examples{
library(SIEVE)
data(clrCounts3) # first 50 genes (gene1 to gene50) are DE genes
groups <- c(rep(0, 200), rep(1, 200))
clrSeq_result3 <- clrSeq(clrCounts3, group = groups) # DE dataset
clrSIEVE_result3 <- clrSIEVE(clrSeq_result = clrSeq_result3,
                             alpha_level = 0.05,
                             order_DE = FALSE,
                             order_LFC = FALSE,
                             order_DS = FALSE,
                             order_sieve = FALSE)
clrDE_test3 <- clrSIEVE_result3$clrDE_test # DE test
head(clrDE_test3, 5)
clrDS_test3 <- clrSIEVE_result3$clrDS_test # DS test
clrDS_test3[clrDS_test3$adj_pval_DS < 0.05, ]
clrSIEVE_tests3 <- clrSIEVE_result3$clrSIEVE_tests # Sieve DE, DV and DS genes
head(clrSIEVE_tests3, 5)
tail(clrSIEVE_tests3, 5)

}
