% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SieveFittingModels.R
\name{create_index_matrix}
\alias{create_index_matrix}
\title{Create the index matrix for multivariate regression}
\usage{
create_index_matrix(xdim, basisN = NULL, maxj = NULL, interaction_order = 5)
}
\arguments{
\item{xdim}{a number. It specifies the predictors' dimension.}

\item{basisN}{a number. The number of basis function to use.}

\item{maxj}{a number. We use this to specify the largest row product in the index list.}

\item{interaction_order}{a number The maximum order of interaction. 1 means additive model, 2 means including pairwise interaction terms, etc.}
}
\value{
a matrix. The first column is the product of the indices, the rest columns are the index vectors for constructing multivariate basis functions.
}
\description{
Create the index matrix for multivariate regression
}
