% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_cells.R
\name{fetch_cells}
\alias{fetch_cells}
\title{Find the overlapped genes of gene sets defined by different method.}
\usage{
fetch_cells(expr,ratio=0.7,n=50)
}
\arguments{
\item{expr}{The expression dataframe. Rows should be genes and columns should be cells.}

\item{ratio}{the Proportion of cell subset.The default value of ratio is set to 0.7, but can also be set to other values between 0 to 1.}

\item{n}{The times of subsampling.The default value of n is set to 50.}
}
\value{
A list of random cell subsets.
}
\description{
Find the overlapped genes of gene sets defined by different method.
}
\examples{
celllist<-fetch_cells(expr,ratio=0.7,n=50)
}
