% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_midpoint}
\alias{find_midpoint}
\title{Find class midpoint}
\usage{
find_midpoint(size)
}
\arguments{
\item{size}{vector of sizes.}
}
\value{
a vector of midpoint sizes.
}
\description{
Find class midpoint
}
