% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multiple_sample}
\alias{multiple_sample}
\title{Multiple sample particle size data}
\format{
\subsection{\code{single_sample}}{

A data frame with 125 rows and 60 columns:
\describe{
\item{size}{Particle size class (µm)}
\item{example_01, example_02, example_03}{Abundance of a particle size class in sample (\%)}
}
}
}
\source{
Maurycy Żarczyński
}
\usage{
multiple_sample
}
\description{
A multiple sample data from particle size analysis
}
\keyword{datasets}
