% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepare_tibble}
\alias{prepare_tibble}
\title{Prepare sieve tibble}
\usage{
prepare_tibble(data, unit)
}
\arguments{
\item{data}{a tibble with size and abundances.}

\item{unit}{original unit of measurements. One of "um" or "phi".}
}
\value{
a tibble prepared for further analyses with \pkg{sieve}.
}
\description{
Prepare sieve tibble
}
