% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_indices.R
\name{psa_calculate_indices}
\alias{psa_calculate_indices}
\title{Calculate base size indices}
\usage{
psa_calculate_indices(data, method = "all")
}
\arguments{
\item{data}{a tibble prepared with \code{\link{prepare_tibble}}}

\item{method}{one of arithmetic method of moments ("amm"), logarithmic method of moments ("lmm") or Geometric method of moments ("gmm") or all (default).}
}
\value{
a data frame with calculated indices
}
\description{
Calculate base size indices
}
