% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_indices.R
\name{psa_calculate_indices_gfw}
\alias{psa_calculate_indices_gfw}
\title{Calculate statistics using Folk and Ward geometric method}
\usage{
psa_calculate_indices_gfw(data)
}
\arguments{
\item{data}{data prepared with \code{\link{psa_prepare_data}}}
}
\value{
a tibble with calculated statistics.
}
\description{
Calculate statistics using Folk and Ward geometric method
}
