% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_indices.R
\name{psa_calculate_indices_gmm}
\alias{psa_calculate_indices_gmm}
\title{Calculate statistics using geometric method of moments}
\usage{
psa_calculate_indices_gmm(data)
}
\arguments{
\item{data}{data prepared with \code{\link{psa_prepare_data}}}
}
\value{
a tibble with calculated statistics.
}
\description{
Calculate statistics using geometric method of moments
}
