% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_indices.R
\name{psa_calculate_indices_lfw}
\alias{psa_calculate_indices_lfw}
\title{Calculate statistics using Folk and Ward logarithmic method}
\usage{
psa_calculate_indices_lfw(data)
}
\arguments{
\item{data}{data prepared with \code{\link{psa_fw_prepare}}}
}
\value{
a tibble with calculated statistics.
}
\description{
Calculate statistics using Folk and Ward logarithmic method
}
