% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_indices.R
\name{psa_calculate_indices_lmm}
\alias{psa_calculate_indices_lmm}
\title{Calculate statistics using logarithmic method of moments}
\usage{
psa_calculate_indices_lmm(data)
}
\arguments{
\item{data}{data prepared with \code{\link{psa_prepare_data}}}
}
\value{
a tibble with calculated statistics.
}
\description{
Calculate statistics using logarithmic method of moments
}
