% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_indices.R
\name{psa_calculate_qunatiles_phi}
\alias{psa_calculate_qunatiles_phi}
\title{Calculate qunatile based phi diameter statistics}
\usage{
psa_calculate_qunatiles_phi(data)
}
\arguments{
\item{data}{data prepared with \code{\link{psa_prepare_data}}}
}
\value{
a tibble with calculated statistics.
}
\description{
Calculate qunatile based phi diameter statistics
}
