% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{psa_convert_units}
\alias{psa_convert_units}
\title{Convert units between metric and phi}
\usage{
psa_convert_units(size, unit)
}
\arguments{
\item{size}{vector (numeric) of sizes.}

\item{unit}{either "µm" or "phi" (character).}
}
\value{
a vector of converted sizes.
}
\description{
Convert units between metric and phi
}
