% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_indices.R
\name{psa_fw_prepare}
\alias{psa_fw_prepare}
\title{Prepare data for Folk and Ward methods}
\usage{
psa_fw_prepare(data)
}
\arguments{
\item{data}{data prepared with \code{\link{psa_prepare_data}}}
}
\value{
a tibble with calculated statistics.
}
\description{
Prepare data for Folk and Ward methods
}
