% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_indices.R
\name{psa_get_indices}
\alias{psa_get_indices}
\title{Get all indices}
\usage{
psa_get_indices(data, method = "all")
}
\arguments{
\item{data}{nested data prepared with \code{\link{prepare_tibble}}}

\item{method}{character, one of: arithmetic method of moments ("amm"), logarithmic method of moments ("lmm"), geometric method of moments ("gmm"), Folk and Ward logarithmic ("lfw"), Folk and Ward geometric ("gfw"), metric quantiles ("umq"), phi quantiles ("phiq") or all ("all" - default).}
}
\value{
a data frame with calculated indices.
}
\description{
Get all indices
}
