% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{psa_plot}
\alias{psa_plot}
\title{Plot variable plot types}
\usage{
psa_plot(data, type, ...)
}
\arguments{
\item{data}{data}

\item{type}{which plot}

\item{...}{optional arguments}
}
\value{
plot.
}
\description{
Plot variable plot types
}
