% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{psa_plot_cumulative}
\alias{psa_plot_cumulative}
\title{Plot cumulative distributions}
\usage{
psa_plot_cumulative(data, unit = "phi")
}
\arguments{
\item{data}{data to plot}

\item{unit}{unit of sizes}
}
\value{
plot(s)
}
\description{
Plot cumulative distributions
}
