% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{psa_plot_distribution}
\alias{psa_plot_distribution}
\title{Plot particle size distributions}
\usage{
psa_plot_distribution(data, unit = "phi")
}
\arguments{
\item{data}{data to plot.}

\item{unit}{unit of sizes.}
}
\value{
particle size distribution plot.
}
\description{
Plot particle size distributions
}
