% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{psa_plot_ternary}
\alias{psa_plot_ternary}
\title{Plot ternary diagrams}
\usage{
psa_plot_ternary(data, triangle)
}
\arguments{
\item{data}{data.}

\item{triangle}{which triangle to use.}
}
\value{
ternary diagram of choice.
}
\description{
Plot ternary diagrams
}
