% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{psa_prepare_data}
\alias{psa_prepare_data}
\title{Prepare and nest particle size data}
\usage{
psa_prepare_data(data, unit = "um")
}
\arguments{
\item{data}{a tibble with size and abundances.}

\item{unit}{original unit of measurements. One of "um" (default) or "phi".}
}
\value{
A nested tibble with data prepared for further analyses with \pkg{psa}.
}
\description{
Add secondary size unit, calculate midpoints and cumulative abundance. Size class must come in either µm or phi.
All abundance data must come as percent (\%).
}
\examples{
psa_prepare_data(multiple_sample, unit = "um")
}
