% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIEVE.R
\name{SIEVE}
\alias{SIEVE}
\title{Find the high frequency genes of gene sets defined by different method.}
\usage{
SIEVE(genelist,n=50)
}
\arguments{
\item{genelist}{The gene sets matrix. Rows should be genes and columns should be different batches.}

\item{n}{The minimum number of repetitions of a gene.}
}
\value{
A list of high frequency genes and a schematic diagram of the gene frequency.
}
\description{
Find the high frequency genes of gene sets defined by different method.
}
\examples{
geneset<-SIEVE(genelist,n=50)

}
