% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{single_sample}
\alias{single_sample}
\title{Single sample particle size data}
\format{
\subsection{\code{single_sample}}{

A data frame with 125 rows and 60 columns:
\describe{
\item{size}{Particle size class (µm)}
\item{example_01}{Abundance of a particle size class in sample (\%)}
}
}
}
\source{
Maurycy Żarczyński
}
\usage{
single_sample
}
\description{
A single sample data from particle size analysis
}
\keyword{datasets}
