% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SN.plot.R
\name{SN.plot}
\alias{SN.plot}
\title{Distribution of CLR-transformed counts}
\usage{
SN.plot(data)
}
\arguments{
\item{data}{CLR-transformed counts for a particular gene/transcript.}
}
\description{
Produce a histogram of observed CLR-transformed counts, with the
fitted skew-normal probability density function for a particular gene/transcript.
}
\examples{
   library(SIEVE)
   data("clrCounts1")
   SN.plot(clrCounts1[1,])
   SN.plot(clrCounts1[2,])
   SN.plot(clrCounts1[3,])

}
