% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violin.plot.SIEVE.R
\name{violin.plot.SIEVE}
\alias{violin.plot.SIEVE}
\title{Violin plots}
\usage{
violin.plot.SIEVE(
  data = NULL,
  name.gene = NULL,
  group = NULL,
  group.names = NULL,
  xlab = "CLR-transformed count",
  ylab = "Condition"
)
}
\arguments{
\item{data}{A CLR-transformed count table.}

\item{name.gene}{Gene/transcript name.}

\item{group}{A vector specifying the group labels of the data.}

\item{group.names}{A vector specifying the group names.}

\item{xlab}{Name of the x-axis.}

\item{ylab}{Name of the y-axis.}
}
\description{
Produce violin plots of CLR-transformed count data for two or three groups.
}
\examples{
  library(SIEVE)
  data(clrCounts2)  # first 50 genes (gene1 to gene50) are DV genes
  data(clrCounts3)  # first 50 genes (gene1 to gene50) are DE genes
  group0 <- c(rep(0, 200), rep(1, 200))
  group1 <- c(rep(0, 200), rep(1, 100), rep(2, 100))
  violin.plot.SIEVE(data = clrCounts2, "gene1", group = group0,
                group.names = c("control", "case")) # DV
  violin.plot.SIEVE(data = clrCounts2, "gene1", group = group1,
                group.names = c("control", "case1", "case2")) # DV
  violin.plot.SIEVE(data = clrCounts3, "gene1", group = group0,
                group.names = c("control", "case")) # DE
  violin.plot.SIEVE(data = clrCounts3, "gene2", group = group0,
                group.names = c("control", "case")) # DE
  violin.plot.SIEVE(data = clrCounts3, "gene200", group = group0,
                group.names = c("control", "case")) # non-DE
}
