% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_indices.R
\name{psa_calculate_indices}
\alias{psa_calculate_indices}
\title{Calculate base size indices}
\usage{
psa_calculate_indices(data, method)
}
\arguments{
\item{data}{a tibble prepared with \code{\link{psa_prepare_data}}}

\item{method}{character, one of: arithmetic method of moments ("amm"), logarithmic method of moments ("lmm"), geometric method of moments ("gmm"), Folk and Ward logarithmic ("lfw"), Folk and Ward geometric ("gfw"), metric quantiles ("umq"), phi quantiles ("phiq") or all ("all" - default).}
}
\value{
a data frame with calculated indices.
}
\description{
Calculate base size indices
}
